refactor: remove V1 renderer, make V2 the default
Remove the old V1 renderer (~3,700 lines) and make the V2 instanced
GPU renderer the default (and only) renderer.

Changes:
- Move shared code from V1 to renderer/ (gpu_context, text, text_measurer)
- Create types.rs for shared types (Gradient, Shadow, etc.)
- Delete V1-only files (primitives.rs, pipeline.rs, shaders)
- Rename renderer_v2/ to renderer/
- Rename RendererV2 to Renderer, PaintContextV2 to PaintContext
- Remove all #[cfg(feature = "renderer_v2")] conditionals
- Rename paint_v2 to paint in Widget trait and all implementations
- Remove unused paint_ctx field from ManagedSurface
- Remove renderer_v2 feature from Cargo.toml
- Update ARCHITECTURE.md with current renderer structure
- Delete obsolete RENDERER_V2_NEXT_STEPS.md
