+++ src/renderer/image_shader.wgsl
