+++ src/surface_manager.rs
@@ -9,14 +9,9 @@ use smithay_client_toolkit::reexports::client::Connection;
 
 use crate::layout::Constraints;
 use crate::platform::{WaylandState, WaylandWindowWrapper};
-use crate::renderer::{GpuContext, PaintContext, SurfaceState};
+use crate::renderer::{GpuContext, SurfaceState};
 use crate::surface::{SurfaceConfig, SurfaceId};
 use crate::widgets::Widget;
-
-// PaintContext capacity constants
-const PAINT_CTX_SHAPES: usize = 32;
-const PAINT_CTX_TEXTS: usize = 16;
-const PAINT_CTX_OVERLAYS: usize = 8;
 
 /// A surface with unified GPU lifecycle management.
 ///
@@ -29,9 +24,6 @@ pub struct ManagedSurface {
     pub config: SurfaceConfig,
     /// The root widget for this surface
     pub widget: Box<dyn Widget>,
-    /// Paint context for rendering (unused when renderer_v2 is enabled)
-    #[cfg_attr(feature = "renderer_v2", allow(dead_code))]
-    pub paint_ctx: PaintContext,
     /// The wgpu surface state (None until GPU init)
     pub wgpu_surface: Option<SurfaceState>,
     /// Previous scale factor for detecting changes
@@ -45,11 +37,6 @@ impl ManagedSurface {
             id,
             config,
             widget,
-            paint_ctx: PaintContext::with_capacity(
-                PAINT_CTX_SHAPES,
-                PAINT_CTX_TEXTS,
-                PAINT_CTX_OVERLAYS,
-            ),
             wgpu_surface: None,
             previous_scale_factor: 1.0,
         }
