+++ src/widgets/container/scrollable.rs
@@ -4,8 +4,6 @@ use crate::animation::{SpringConfig, Transition};
 use crate::layout::Constraints;
 use crate::reactive::request_animation_frame;
 use crate::renderer::PaintContext;
-#[cfg(feature = "renderer_v2")]
-use crate::renderer_v2::PaintContextV2;
 use crate::transform::Transform;
 use crate::transform_origin::TransformOrigin;
 use crate::widgets::scroll::{ScrollAxis, ScrollbarAxis, ScrollbarVisibility};
@@ -372,34 +370,6 @@ impl Container {
 
     /// Paint scrollbar container widgets
     pub(super) fn paint_scrollbar_containers(&self, ctx: &mut PaintContext) {
-        if self.scrollbar_visibility == ScrollbarVisibility::Hidden {
-            return;
-        }
-
-        // Vertical scrollbar
-        if self.scroll_axis.allows_vertical() && self.scroll_state.needs_vertical_scrollbar() {
-            if let Some(ref track) = self.v_scrollbar_track {
-                track.paint(ctx);
-            }
-            if let Some(ref handle) = self.v_scrollbar_handle {
-                handle.paint(ctx);
-            }
-        }
-
-        // Horizontal scrollbar
-        if self.scroll_axis.allows_horizontal() && self.scroll_state.needs_horizontal_scrollbar() {
-            if let Some(ref track) = self.h_scrollbar_track {
-                track.paint(ctx);
-            }
-            if let Some(ref handle) = self.h_scrollbar_handle {
-                handle.paint(ctx);
-            }
-        }
-    }
-
-    /// Paint scrollbar container widgets (V2 renderer)
-    #[cfg(feature = "renderer_v2")]
-    pub(super) fn paint_scrollbar_containers_v2(&self, ctx: &mut PaintContextV2) {
         use crate::transform::Transform;
         use crate::widgets::Rect;
 
@@ -418,7 +388,7 @@ impl Container {
 
                 let mut track_ctx = ctx.add_child(0, track_local);
                 track_ctx.set_transform(Transform::translate(track_offset_x, track_offset_y));
-                track.paint_v2(&mut track_ctx);
+                track.paint(&mut track_ctx);
             }
             if let Some(ref handle) = self.v_scrollbar_handle {
                 let handle_global = handle.bounds();
@@ -428,7 +398,7 @@ impl Container {
 
                 let mut handle_ctx = ctx.add_child(0, handle_local);
                 handle_ctx.set_transform(Transform::translate(handle_offset_x, handle_offset_y));
-                handle.paint_v2(&mut handle_ctx);
+                handle.paint(&mut handle_ctx);
             }
         }
 
@@ -442,7 +412,7 @@ impl Container {
 
                 let mut track_ctx = ctx.add_child(0, track_local);
                 track_ctx.set_transform(Transform::translate(track_offset_x, track_offset_y));
-                track.paint_v2(&mut track_ctx);
+                track.paint(&mut track_ctx);
             }
             if let Some(ref handle) = self.h_scrollbar_handle {
                 let handle_global = handle.bounds();
@@ -452,7 +422,7 @@ impl Container {
 
                 let mut handle_ctx = ctx.add_child(0, handle_local);
                 handle_ctx.set_transform(Transform::translate(handle_offset_x, handle_offset_y));
-                handle.paint_v2(&mut handle_ctx);
+                handle.paint(&mut handle_ctx);
             }
         }
     }
