+++ src/widgets/text_input.rs
@@ -19,8 +19,6 @@ use crate::reactive::{
     request_animation_frame, request_focus, set_cursor, start_layout_tracking,
 };
 use crate::renderer::{PaintContext, char_index_from_x_styled, measure_text_styled};
-#[cfg(feature = "renderer_v2")]
-use crate::renderer_v2::PaintContextV2;
 
 use super::font::{FontFamily, FontWeight};
 use super::widget::{Color, Event, EventResponse, Key, Modifiers, MouseButton, Rect, Widget};
@@ -1032,64 +1030,6 @@ impl Widget for TextInput {
     }
 
     fn paint(&self, ctx: &mut PaintContext) {
-        let display = self.display_text_cached();
-        let text_color = self.text_color.get();
-        let is_focused = has_focus(self.widget_id);
-
-        // Clip to bounds (prevents text overflow)
-        ctx.push_clip(self.bounds, 0.0, 1.0);
-
-        // Draw selection highlight if focused and has selection
-        if is_focused && self.selection.has_selection() {
-            let (start, end) = self.selection.range();
-            // Use cached glyph positions instead of measuring
-            let start_x = self.cached_width_at_char(start) - self.scroll_offset;
-            let end_x = self.cached_width_at_char(end) - self.scroll_offset;
-
-            let selection_rect = Rect::new(
-                self.bounds.x + start_x,
-                self.bounds.y,
-                end_x - start_x,
-                self.bounds.height,
-            );
-            ctx.draw_rect(selection_rect, self.selection_color.get());
-        }
-
-        // Draw text with scroll offset
-        // Use cached text width to prevent word wrapping - the clip rect handles clipping
-        let text_bounds = Rect::new(
-            self.bounds.x - self.scroll_offset,
-            self.bounds.y,
-            self.cached_text_width.max(self.bounds.width),
-            self.bounds.height,
-        );
-        ctx.draw_text_styled(
-            display,
-            text_bounds,
-            text_color,
-            self.cached_font_size,
-            self.cached_font_family.clone(),
-            self.cached_font_weight,
-        );
-
-        // Draw cursor if focused and visible
-        if is_focused && self.cursor_visible {
-            // Use cached glyph position
-            let cursor_x = self.cached_width_at_char(self.selection.cursor) - self.scroll_offset;
-            let cursor_rect = Rect::new(
-                self.bounds.x + cursor_x,
-                self.bounds.y,
-                1.5, // cursor width
-                self.bounds.height,
-            );
-            ctx.draw_rect(cursor_rect, self.cursor_color.get());
-        }
-
-        ctx.pop_clip();
-    }
-
-    #[cfg(feature = "renderer_v2")]
-    fn paint_v2(&self, ctx: &mut PaintContextV2) {
         // Draw in LOCAL coordinates (0,0 is widget origin)
         // Parent Container sets position transform
         let display = self.display_text_cached();
